% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attrition.R
\name{attrition}
\alias{attrition}
\title{Quickly get an attrition plot}
\usage{
attrition(inputs, losses, unit, labels, format = "plot")
}
\arguments{
\item{inputs}{Numeric vector of n's going in}

\item{losses}{Numeric vector of n's going out - last is those included}

\item{unit}{Preface a unit?}

\item{labels}{Character vector of labels - in order of inputs then losses.}

\item{format}{The default is "plot," which produces a plot in the R graphics device. Current alternate options include "pdf" and "bmp," which produce those file types under the name "Sankey.xxx" in the current directory.}
}
\description{
This function will take raw values. It's based on a stackoverflow post (see raw code).
}
\section{Bugs}{

If you find an issue, post it here and we'll try and fix it: \url{http://go.roche.com/RochePlotissue}
}

\examples{
attrition(
  inputs = c(212,304),
  losses = c(111,134,271),
  unit = "",
  labels = c("Medicare","Commercial","No followup","Missing something","Included")
)
}
\keyword{attrition}
\keyword{sankey}
\keyword{teradata}
