% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggAttrition.R
\name{ggAttrition}
\alias{ggAttrition}
\title{Visualise attrition data as a barplot}
\usage{
ggAttrition(data, label, n, sort = NULL, show_p = TRUE, N_max = NULL,
  bar_fill = NULL, bar_col = "black", bar_alpha = 0.9,
  label_pos = "b", label_thresh = 0.7, title_pos = "m",
  wrap_width = 100, flip = TRUE, nudge_title = 10,
  nudge_lab = NULL)
}
\arguments{
\item{data}{dataframe.}

\item{label}{Unquoted name of column indicating selection criteria}

\item{n}{Unquoted name of column indicating cohort size}

\item{sort}{Either NULL (no sorting), 'a' for ascending, or 'd' for descending}

\item{show_p}{logical. Should percentages be included?}

\item{N_max}{numeric. Maximum value used in calculating percentages. Defaults to the maximum value observed in the data.}

\item{bar_fill}{character. Optional bar fill and border color}

\item{bar_col}{character. Optional bar border color}

\item{bar_alpha}{numeric. Alpha of bar}

\item{label_pos}{character. Position of labels, either "b", "m", or "l"}

\item{title_pos}{character. Position of title, either "m" for margins, or "l" for left side of plot}

\item{wrap_width}{integer. Optional integer indicating how much to wrap selection criteria text}

\item{flip}{logical. Should axes be flipped?}

\item{nudge_title}{numeric. Nudging of title}

\item{nudge_lab}{numeric. Nudging of label}

\item{type}{character. Type of plot. Currently must be 'bar'}

\item{label_pos}{character. Position of labels, either "t" for top, or "b" for bottom.}
}
\description{
Visualise attrition data as a barplot
}
\examples{

 library(tidyverse)

 data <- tibble("criteria" = c("Start",
                               "Criteria A ....",
                               "Criteria B ....",
                               "Criteria C ....",
                               "Criteria D ...."),
                "size" = c(1000, 785, 456, 234, 154))


ggAttrition(data, criteria, size) +
  labs(title = "Attrition Table")


}
