% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sunburst.R
\name{sunburst}
\alias{sunburst}
\title{Quickly get a D3 js sunburst}
\usage{
sunburst(dataframe, id, linename, linenumber, n_common = 5,
  return_data = F, legend_width = 300)
}
\arguments{
\item{dataframe}{the dataframe holding the data}

\item{id}{grouping variable that links the long data. Most likely patient id.}

\item{linename}{The name of the line - or whatever goes in the bars}

\item{linenumber}{A column with the numeric line number.}

\item{n_common}{The n most common treatments you want to include. Others will go into no treatment or other.}

\item{return_data}{FALSE by default. If true, it will return the formatted data for you to pass to the plotting function yourself.}

\item{legend_width}{Width of the legend in pixels. This will likely need tweaking.}
}
\description{
This function will take 4 columns of data, format the data, then feed it to sunburstR where it will be converted into a D3 plot.
}
\section{Bugs}{

If you find an issue, post it here and we'll try and fix it: \url{http://go.roche.com/RochePlotissue}
}

\examples{
\dontrun{
          ##### Get data from flatiron ##################
             library(dplyr)


             # data valid today
             valid_date <- Sys.Date()

             # If on Windows, life is no fun
             if (Sys.info()[["sysname"]] == "Windows"){
               Sys.setenv(JAVA_HOME="")  # for java - go to 64bit
             }

             # connect to data
             tdRWDSconnect(
               datalab = FALSE,
               type = "teradataR"
             )

             line.of.therapy <- tdRWDSquery(paste0(
               "
               SELECT
               patientid
               ,LineName
               ,LineNumber
               ,IsMaintenanceTherapy
               FROM rwd_vdm_flatiron.v_cll_lineoftherapy
               WHERE
               VALID_START <= CAST('",paste0(valid_date," 00:00:00"),"' AS TIMESTAMP) AND
               VALID_END >= CAST('",paste0(valid_date," 00:00:00"),"' AS TIMESTAMP)
               ;"
             )) \%>\%
               filter(
                 # just do lines 0 to 3
                 LineNumber < 4 &
                   # and ignore maintenence
                   IsMaintenanceTherapy == "False"
               )

             ##### Use function ##################

             sunburst(
               dataframe = line.of.therapy,
               id = "PatientID",
               linename = "LineName",
               linenumber = "LineNumber",
               n_common = 3,
               legend_width = 300
             )

          }
}
\keyword{sunburst}
\keyword{teradata}
