% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sankey.R
\name{sankey}
\alias{sankey}
\title{Quickly get a D3 js sankey}
\usage{
sankey(dataframe, id = NULL, linename = NULL, linenumber = NULL,
  n_common = 3, return_data = F, postfix = " patients")
}
\arguments{
\item{dataframe}{the dataframe holding the data}

\item{id}{grouping variable that links the long data. Most likely patient id.}

\item{linename}{The name of the line - or whatever goes in the bars}

\item{linenumber}{A column with the numeric line number.}

\item{n_common}{The n most common treatments you want to include. Others will go into no treatment or other.}

\item{return_data}{FALSE by default. If true, it will return the formatted data for you to pass to the plotting function yourself.}

\item{postfix}{What do you want to go on the labels for the Sankey}
}
\description{
This function will take 4 columns of data, format the data, then feed it to networkD3 where it will be converted into a D3 plot.
}
\section{Vignette}{

This function has a vignette - \url{https://pages.github.roche.com/Rpackages/RochePlot/articles/pkgdown_only/sankey.html}
}

\section{Bugs}{

If you find an issue, post it here and we'll try and fix it: \url{http://go.roche.com/RochePlotissue}
}

\examples{
paste("See https://pages.github.roche.com/Rpackages/RochePlot/articles/pkgdown_only/sankey.html")
}
\keyword{sankey}
\keyword{teradata}
